Rename chalice and fastapi folders for improved structure
