+++ examples/chalice/__init__.py
