+++ examples/chalice/blueprints/authed.py
