+++ examples/chalice/models/files.py
