+++ examples/chalice/resources/__init__.py
