+++ examples/chalice/resources/accounts.py
