+++ examples/chalice/resources/base.py
