+++ examples/chalice/resources/files.py
