+++ examples/chalice/resources/users.py
