+++ examples/fastapi/blueprints/custom_query_blueprint.py
