+++ examples/fastapi/middlewares/authed.py
