+++ examples/fastapi/models/transactions.py
