+++ examples/fastapi/models/users.py
