+++ examples/fastapi/resources/__init__.py
