+++ examples/fastapi/resources/accounts.py
