+++ examples/fastapi/resources/base.py
