+++ examples/fastapi/resources/cards.py
