+++ examples/fastapi/resources/files.py
