+++ examples/fastapi/resources/transactions.py
