+++ examples/fastapi/resources/users.py
