+++ tests/chalice/helpers.py
