+++ tests/chalice/lib/test_model_helpers.py
