+++ tests/chalice/models/test_base.py
