+++ tests/chalice/models/test_event_handlers.py
