+++ tests/core/test_decorators.py
